# ci(integration-cli): split test suites in a matrix