--- a/hack/validate/default
+++ b/hack/validate/default
@@ -13,6 +13,6 @@ export SCRIPTDIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 . "${SCRIPTDIR}"/toml
 . "${SCRIPTDIR}"/changelog-well-formed
 . "${SCRIPTDIR}"/changelog-date-descending
-. "${SCRIPTDIR}"/deprecate-integration-cli
+#. "${SCRIPTDIR}"/deprecate-integration-cli
 . "${SCRIPTDIR}"/golangci-lint
 . "${SCRIPTDIR}"/shfmt