--- a/integration-cli/check_test.go
+++ b/integration-cli/check_test.go
@@ -83,7 +83,240 @@ func ensureTestEnvSetup(t *testing.T) {
 
 func TestDockerSuite(t *testing.T) {
 	ensureTestEnvSetup(t)
-	suite.Run(t, &DockerSuite{})
+	suite.Run(t, &DockerAPISuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerBenchmarkSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIAttachSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIBuildSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLICommitSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLICpSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLICreateSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIEventSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIExecSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIHealthSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIHistorySuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIImagesSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIImportSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIInfoSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIInspectSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLILinksSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLILoginSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLILogsSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLINetmodeSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLINetworkSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIPluginLogDriverSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIPluginsSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIPortSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIProxySuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIPruneSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIPsSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIPullSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIPushSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIRestartSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIRmiSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIRunSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLISaveLoadSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLISearchSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLISNISuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIStartSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIStatsSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLITopSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIUpdateSuite{ds: &DockerSuite{}})
+	suite.Run(t, &DockerCLIVolumeSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerAPISuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerAPISuite{ds: &DockerSuite{}})
+}
+
+func TestDockerBenchmarkSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerBenchmarkSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIAttachSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIAttachSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIBuildSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIBuildSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLICommitSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLICommitSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLICpSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLICpSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLICreateSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLICreateSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIEventSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIEventSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIExecSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIExecSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIHealthSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIHealthSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIHistorySuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIHistorySuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIImagesSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIImagesSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIImportSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIImportSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIInfoSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIInfoSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIInspectSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIInspectSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLILinksSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLILinksSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLILoginSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLILoginSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLILogsSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLILogsSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLINetmodeSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLINetmodeSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLINetworkSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLINetworkSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIPluginLogDriverSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIPluginLogDriverSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIPluginsSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIPluginsSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIPortSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIPortSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIProxySuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIProxySuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIPruneSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIPruneSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIPsSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIPsSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIPullSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIPullSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIPushSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIPushSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIRestartSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIRestartSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIRmiSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIRmiSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIRunSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIRunSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLISaveLoadSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLISaveLoadSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLISearchSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLISearchSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLISNISuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLISNISuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIStartSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIStartSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIStatsSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIStatsSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLITopSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLITopSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIUpdateSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIUpdateSuite{ds: &DockerSuite{}})
+}
+
+func TestDockerCLIVolumeSuite(t *testing.T) {
+	ensureTestEnvSetup(t)
+	suite.Run(t, &DockerCLIVolumeSuite{ds: &DockerSuite{}})
 }
 
 func TestDockerRegistrySuite(t *testing.T) {