--- a/integration-cli/docker_cli_prune_test.go
+++ b/integration-cli/docker_cli_prune_test.go
@@ -1,0 +1,5 @@
+package main
+
+type DockerCLIPruneSuite struct {
+	ds *DockerSuite
+}