--- a/integration-cli/docker_cli_update_test.go
+++ b/integration-cli/docker_cli_update_test.go
@@ -1,0 +1,5 @@
+package main
+
+type DockerCLIUpdateSuite struct {
+	ds *DockerSuite
+}