Split `DockerSuite` for `integration-cli` tests so we can create a matrix to reduce build time in our pipeline (windows atm). We could also use `tags` for this purpose. Let me know what sounds better to you. Keeping the working tree like this for visibility but will squash commits if it sgty.

Here is the overall build time for each of them after splitting (still windows):

```
- DockerApiSuite: ~12m06s
- DockerBenchmarkSuite: ~11s
- DockerCliAttachSuite: ~16s
- DockerCliBuildSuite: ~16m33s
- DockerCliCommitSuite: ~33s
- DockerCliCpSuite: ~54s
- DockerCliCreateSuite: ~45s
- DockerCliEventSuite: ~1m31s
- DockerCliExecSuite: ~26s
- DockerCliHealthSuite: ~8s
- DockerCliHistorySuite: ~32s
- DockerCliImagesSuite: ~27s
- DockerCliImportSuite: ~8s
- DockerCliInfoSuite: ~8s
- DockerCliInspectSuite: ~52s
- DockerCliLinksSuite: ~9s
- DockerCliLoginSuite: ~8s
- DockerCliLogsSuite: ~48s
- DockerCliNetmodeSuite: ~8s
- DockerCliNetworkSuite: ~8s
- DockerCliPluginLogDriverSuite: ~8s
- DockerCliPluginsSuite: ~10s
- DockerCliPortSuite: ~11s
- DockerCliProxySuite: ~8s
- DockerCliPruneSuite: ~9s
- DockerCliPsSuite: ~1m19s
- DockerCliPullSuite: ~9s
- DockerCliPushSuite: ~13s
- DockerCliRestartSuite: ~58s
- DockerCliRmiSuite: ~1m52s
- DockerCliRunSuite: ~4m07s
- DockerCliSaveLoadSuite: ~8s
- DockerCliSearchSuite: ~11s
- DockerCliSniSuite: ~10s
- DockerCliStartSuite: ~26s
- DockerCliStatsSuite: ~7s
- DockerCliTopSuite: ~16s
- DockerCliUpdateSuite: ~11s
- DockerNetworkSuite|DockerHubPullSuite|DockerRegistrySuite|DockerSchema1RegistrySuite|DockerRegistryAuthTokenSuite|DockerRegistryAuthHtpasswdSuite: ~11s
- DockerSwarmSuite: ~8s
- DockerDaemonSuite: ~8s
- DockerExternalVolumeSuite: ~7s
```

Before ~1h20m

After ~42m

![image](https://user-images.githubusercontent.com/1951866/171687129-9ee7b1c2-ce0d-413c-a119-6b853f7fb06b.png)