@crazy-max 2022-06-02
@thaJeztah Not linked to these changes but looks like we have an issue with criu repo: https://ci-next.docker.com/public/blue/organizations/jenkins/moby/detail/PR-43682/1/pipeline#step-163-log-428

```
[2022-06-02T17:20:49.208Z] #51 ERROR: executor failed running [/bin/sh -c echo 'deb https://download.opensuse.org/repositories/devel:/tools:/criu/Debian_11/ /' > /etc/apt/sources.list.d/criu.list         && apt-get update         && apt-get install -y --no-install-recommends criu         && install -D /usr/sbin/criu /build/criu]: exit code: 100
```