@crazy-max 2022-06-02
Force pushed and looks worse now :sweat_smile::

```
[2022-06-02T17:57:17.080Z] #51 19.97 E: Failed to fetch https://mirrorcache-us.opensuse.org/repositories/devel:/tools:/criu/Debian_11/amd64/criu_3.17-1_amd64.deb  Could not connect to mirrorcache-us.opensuse.org:443 (91.193.113.65). - connect (111: Connection refused) Could not connect to mirrorcache-us.opensuse.org:443 (2a07:de40:401::65). - connect (101: Network is unreachable)
```