@tianon 2022-06-03
Very interesting!  I like the idea, but I'm concerned about how repetitive it is (both in the code and in the configuration) -- for example, maintaining that long list of tiny functions and/or the list of all their names in the YAML seems really error prone. :grimacing: