@crazy-max 2022-06-03
> Very interesting! I like the idea, but I'm concerned about how repetitive it is (both in the code and in the configuration) -- for example, maintaining that long list of tiny functions and/or the list of all their names in the YAML seems really error prone. 😬

Yes I'm also worried about missing some tests in the future using this pattern. I will check if go tooling allows to have some stats with available tests and suites in a pkg and split them accordingly in the matrix workflow like [this](https://github.com/crazy-max/log4j-cve-scan/blob/4f5af08475dd1486c00f0e24202998aa54f533dc/.github/workflows/dvp.yml#L32).