@crazy-max 2022-06-03
@thaJeztah Looks like we are not able to update or add new tests in integration-cli: https://ci-next.docker.com/public/blue/organizations/jenkins/moby/detail/PR-43682/7/pipeline#step-256-log-7 :sweat_smile:

```
[2022-06-03T10:34:23.854Z] The following new tests were added to integration-cli:
[2022-06-03T10:34:23.854Z] 
[2022-06-03T10:34:23.854Z] +func (s *DockerApiSuite) TestGetContainersAttachWebsocket(c *testing.T) {
[2022-06-03T10:34:23.854Z] +func (s *DockerApiSuite) TestPostContainersAttachContainerNotFound(c *testing.T) {
[2022-06-03T10:34:23.854Z] +func (s *DockerApiSuite) TestGetContainersWsAttachContainerNotFound(c *testing.T) {
[2022-06-03T10:34:23.854Z] +func (s *DockerApiSuite) TestPostContainersAttach(c *testing.T) {
```