@crazy-max 2022-06-04
To dynamically generate and distribute tests in a matrix I checked [`go test -list`](https://github.com/golang/go/blob/4aa1efed4853ea067d665a952eee77c52faac774/src/cmd/go/internal/test/test.go#L278-L282) but unfortunately it doesn't list test suites.

I have created a small library called [gotestlist](https://github.com/crazy-max/gotestlist) to list tests in the given Go packages. Here is what it looks if I run `gotestlist` in `integration-cli`:

```console
$ go install github.com/crazy-max/gotestlist/cmd/gotestlist@latest
$ cd integration-cli/
$ gotestlist -d 4 ./... | jq
```
```json
[
  "DockerAPISuite|DockerBenchmarkSuite|DockerCliAttachSuite",
  "DockerCliBuildSuite|DockerCliCommitSuite|DockerCliCpSuite|DockerCliCreateSuite|DockerCliEventSuite",
  "DockerCliExecSuite|DockerCliHealthSuite|DockerCliHistorySuite|DockerCliImagesSuite|DockerCliImportSuite|DockerCliInfoSuite|DockerCliInspectSuite|DockerCliLinksSuite|DockerCliLoginSuite|DockerCliLogsSuite|DockerCliNetmodeSuite|DockerCliNetworkSuite|DockerCliPluginLogDriverSuite|DockerCliPluginsSuite|DockerCliPortSuite|DockerCliProxySuite|DockerCliPruneSuite|DockerCliPsSuite|DockerCliPullSuite|DockerCliPushSuite|DockerCliRestartSuite|DockerCliRmiSuite",
  "DockerCliRunSuite|DockerCliSaveLoadSuite|DockerCliSearchSuite|DockerCliSniSuite|DockerCliStartSuite|DockerCliStatsSuite|DockerCliTopSuite|DockerCliUpdateSuite|DockerCliVolumeSuite|DockerDaemonSuite|DockerExternalVolumeSuite|DockerHubPullSuite|DockerNetworkSuite|DockerPluginSuite|DockerRegistryAuthHtpasswdSuite|DockerRegistryAuthTokenSuite|DockerRegistrySuite|DockerSchema1RegistrySuite|DockerSwarmSuite"
]
```

`-d` flag dynamically distributes the tests based on the given matrix size (here 4). This JSON output can then be used as matrix input in our workflow. This way no tests are missing.

I don't calculate cyclomatic complexities of tests functions in gotestlist to add some "weight" and distribute a bit better but I think it's ok in the meantime.