@crazy-max 2022-06-06
@corhere @neersighted @thaJeztah As discussed on Slack, I look at how we could aggregate data from JSON reports of Go tests. Last commit adds a step for unit and integration tests that will render the output as markdown and create a GitHub summary in our workflow:

![image](https://user-images.githubusercontent.com/1951866/172185653-9721ded8-e6de-41e5-9dc9-256b1a6a0577.png)

There are also some useful stats about slow tests:

![image](https://user-images.githubusercontent.com/1951866/172185829-8efcb094-e822-4f05-8e17-f73569a47584.png)

Summary also includes failed tests with logs:

![image](https://user-images.githubusercontent.com/1951866/172186027-e3a4fcaf-2952-4897-8a26-2aa4f635f155.png)