@neersighted 2022-06-06
> @corhere @neersighted @thaJeztah As discussed on Slack, I looked at how we could aggregate data from JSON reports of Go tests. Last commit adds a step for unit and integration tests that will render the output as markdown and create a GitHub summary in our workflow:
> ...

Nice! Looks like teststat was a solid find -- everything here looks good to me as long as there's no interest in templating/generating the workflows to avoid duplication.