@tianon 2022-06-06
Nice, `gotestlist` is really clever -- I'm assuming the lists have such dramatically different numbers of suites due to the number of tests in each suite, right? (Just to make sure I'm understanding the implementation correctly :sweat_smile:)