@crazy-max 2022-06-17
> We should also have a look at (finally) renaming all the `c *testing.T` to `t *testing.T`, but that's gonna be more fun to do (there's some tests that will have a conflicting variable, so may not be a straight "rename").

Yes let's do that in a follow-up