@thaJeztah 2022-06-22
I'm bringing this one in; there was some discussion on Slack about the large number of jobs (as there's a job limit in GitHub actions); we'll likely run into that limit if we're gonna add Linux as well, but I think this gives the foundation to be flexible, and we can re-slice-and-dice once we run into those limits